requirements.txt
finance-datareader
pykrx
pandas
numpy
requests
plotly
python-dotenv
pyyaml
